refactor(gateway): extract route_responses orchestration into responses/route.rs
Move ~145 lines of inline orchestration logic from the RouterTrait
impl in router.rs into a dedicated responses/route.rs module,
mirroring the existing chat.rs delegation pattern.

What changed:
- model_gateway/src/routers/openai/responses/route.rs: new file with
  ResponsesRouterContext struct and route_responses() free function
  containing all responses orchestration logic (metrics, worker
  selection, validation, history loading, payload prep, provider
  transform, RequestContext setup, streaming dispatch, duration metrics)
- model_gateway/src/routers/openai/responses/mod.rs: add pub(crate)
  mod route declaration
- model_gateway/src/routers/openai/router.rs: replace 145-line inline
  route_responses with 8-line delegation (pack deps + delegate),
  remove dead responses_components() helper, clean up unused imports
  (Instant, to_value, ResponseInput, ResponseInputOutputItem,
  ComponentRefs, PayloadState, RequestContext, WorkerSelection,
  handle_non_streaming_response, handle_streaming_response, Endpoint,
  bool_to_static_str, error)
- model_gateway/src/routers/openai/chat.rs: rename RouterContext to
  ChatRouterContext for consistency with ResponsesRouterContext

Why: router.rs now has both route_chat and route_responses following
the same delegation pattern — pack borrowed refs into a context struct,
delegate to a module-level free function. This reduces router.rs from
~422 to ~278 lines of pure routing dispatch.

How: pure mechanical extraction — the orchestration logic is moved
verbatim with self.* references replaced by deps.* accesses on the
new context struct. Worker selection is inlined via WorkerSelector::new()
matching the chat.rs pattern.
