+++ model_gateway/src/routers/openai/router.rs
@@ -1,7 +1,6 @@
 use std::{
     any::Any,
     sync::{atomic::AtomicBool, Arc},
-    time::Instant,
 };
 
 use axum::{body::Body, extract::Request, http::HeaderMap, response::Response};
@@ -11,27 +10,22 @@ use openai_protocol::{
         RealtimeClientSecretCreateRequest, RealtimeSessionCreateRequest,
         RealtimeTranscriptionSessionCreateRequest,
     },
-    responses::{ResponseInput, ResponseInputOutputItem, ResponsesGetParams, ResponsesRequest},
+    responses::{ResponsesGetParams, ResponsesRequest},
 };
-use serde_json::to_value;
 
 use super::{
-    chat::{self, RouterContext},
-    context::{
-        ComponentRefs, PayloadState, RequestContext, ResponsesComponents, SharedComponents,
-        WorkerSelection,
-    },
+    chat::{self, ChatRouterContext},
+    context::{ResponsesComponents, SharedComponents},
     health,
     provider::ProviderRegistry,
-    responses::{handle_non_streaming_response, handle_streaming_response},
+    responses::route::{self as responses_route, ResponsesRouterContext},
 };
 use crate::{
     app_context::AppContext,
     config::types::RetryConfig,
-    core::{Endpoint, ProviderType, Worker, WorkerRegistry},
-    observability::metrics::{bool_to_static_str, metrics_labels, Metrics},
+    core::{ProviderType, Worker, WorkerRegistry},
+    observability::metrics::{metrics_labels, Metrics},
     routers::{
-        error,
         header_utils::extract_auth_header,
         openai::realtime::{rest::forward_realtime_rest, ws::handle_realtime_ws, RealtimeRegistry},
         worker_selection::{SelectWorkerRequest, WorkerSelector},
@@ -130,10 +124,6 @@ impl OpenAIRouter {
             })
             .await
     }
-
-    fn responses_components(&self) -> Arc<ResponsesComponents> {
-        Arc::clone(&self.responses_components)
-    }
 }
 
 #[async_trait::async_trait]
@@ -156,7 +146,7 @@ impl crate::routers::RouterTrait for OpenAIRouter {
         body: &ChatCompletionRequest,
         model_id: Option<&str>,
     ) -> Response {
-        let deps = RouterContext {
+        let deps = ChatRouterContext {
             worker_registry: &self.worker_registry,
             provider_registry: &self.provider_registry,
             shared_components: &self.shared_components,
@@ -172,146 +162,13 @@ impl crate::routers::RouterTrait for OpenAIRouter {
         body: &ResponsesRequest,
         model_id: Option<&str>,
     ) -> Response {
-        let start = Instant::now();
-        let model = model_id.unwrap_or(body.model.as_str());
-        let streaming = body.stream.unwrap_or(false);
-
-        Metrics::record_router_request(
-            metrics_labels::ROUTER_OPENAI,
-            metrics_labels::BACKEND_EXTERNAL,
-            metrics_labels::CONNECTION_HTTP,
-            model,
-            metrics_labels::ENDPOINT_RESPONSES,
-            bool_to_static_str(streaming),
-        );
-
-        let worker = match self.select_worker(model, headers).await {
-            Ok(w) => w,
-            Err(response) => {
-                Metrics::record_router_error(
-                    metrics_labels::ROUTER_OPENAI,
-                    metrics_labels::BACKEND_EXTERNAL,
-                    metrics_labels::CONNECTION_HTTP,
-                    model,
-                    metrics_labels::ENDPOINT_RESPONSES,
-                    metrics_labels::ERROR_NO_WORKERS,
-                );
-                return response;
-            }
+        let deps = ResponsesRouterContext {
+            worker_registry: &self.worker_registry,
+            provider_registry: &self.provider_registry,
+            responses_components: &self.responses_components,
+            client: &self.responses_components.shared.client,
         };
-
-        // Validate mutual exclusivity of conversation and previous_response_id
-        // Treat empty strings as unset to match other metadata paths
-        let has_conversation = body.conversation.as_ref().is_some_and(|s| !s.is_empty());
-        let has_previous_response = body
-            .previous_response_id
-            .as_ref()
-            .is_some_and(|s| !s.is_empty());
-        if has_conversation && has_previous_response {
-            Metrics::record_router_error(
-                metrics_labels::ROUTER_OPENAI,
-                metrics_labels::BACKEND_EXTERNAL,
-                metrics_labels::CONNECTION_HTTP,
-                model,
-                metrics_labels::ENDPOINT_RESPONSES,
-                metrics_labels::ERROR_VALIDATION,
-            );
-            return error::bad_request(
-                "invalid_request",
-                "Cannot specify both 'conversation' and 'previous_response_id'".to_string(),
-            );
-        }
-
-        let mut request_body = body.clone();
-        if let Some(model) = model_id {
-            request_body.model = model.to_string();
-        }
-        request_body.conversation = None;
-
-        let original_previous_response_id = match super::responses::history::load_input_history(
-            &self.responses_components,
-            body,
-            &mut request_body,
-            model,
-        )
-        .await
-        {
-            Ok(id) => id,
-            Err(response) => return response,
-        };
-
-        request_body.store = Some(false);
-        if let ResponseInput::Items(ref mut items) = request_body.input {
-            items.retain(|item| !matches!(item, ResponseInputOutputItem::Reasoning { .. }));
-        }
-
-        let mut payload = match to_value(&request_body) {
-            Ok(v) => v,
-            Err(e) => {
-                Metrics::record_router_error(
-                    metrics_labels::ROUTER_OPENAI,
-                    metrics_labels::BACKEND_EXTERNAL,
-                    metrics_labels::CONNECTION_HTTP,
-                    model,
-                    metrics_labels::ENDPOINT_RESPONSES,
-                    metrics_labels::ERROR_VALIDATION,
-                );
-                return error::bad_request(
-                    "invalid_request",
-                    format!("Failed to serialize request: {e}"),
-                );
-            }
-        };
-
-        let provider = resolve_provider(&self.provider_registry, worker.as_ref(), model);
-        if let Err(e) = provider.transform_request(&mut payload, Endpoint::Responses) {
-            Metrics::record_router_error(
-                metrics_labels::ROUTER_OPENAI,
-                metrics_labels::BACKEND_EXTERNAL,
-                metrics_labels::CONNECTION_HTTP,
-                model,
-                metrics_labels::ENDPOINT_RESPONSES,
-                metrics_labels::ERROR_VALIDATION,
-            );
-            return error::bad_request("invalid_request", format!("Provider transform error: {e}"));
-        }
-
-        let mut ctx = RequestContext::for_responses(
-            Arc::new(body.clone()),
-            headers.cloned(),
-            model_id.map(String::from),
-            ComponentRefs::Responses(self.responses_components()),
-        );
-
-        ctx.state.worker = Some(WorkerSelection {
-            worker: Arc::clone(&worker),
-            provider: Arc::clone(&provider),
-        });
-
-        ctx.state.payload = Some(PayloadState {
-            json: payload,
-            url: format!("{}/v1/responses", worker.url()),
-            previous_response_id: original_previous_response_id,
-        });
-
-        let response = if ctx.is_streaming() {
-            handle_streaming_response(ctx).await
-        } else {
-            handle_non_streaming_response(ctx).await
-        };
-
-        if response.status().is_success() {
-            Metrics::record_router_duration(
-                metrics_labels::ROUTER_OPENAI,
-                metrics_labels::BACKEND_EXTERNAL,
-                metrics_labels::CONNECTION_HTTP,
-                model,
-                metrics_labels::ENDPOINT_RESPONSES,
-                start.elapsed(),
-            );
-        }
-
-        response
+        responses_route::route_responses(&deps, headers, body, model_id).await
     }
 
     async fn get_response(
